Moved 2 functions from `lattice._lattice` into `lattice.utils`
